Done add new playlist, delete playlist from database Can click playlist to show the list songs screen of album

+++ app/src/main/java/com/dev/musicplayer/data/local/entities/Playlist.kt
@@ -10,5 +10,5 @@ data class Playlist(
     val title: String,
     val thumbnail: String? = null, // image url
     val songs: List<String>? = null, //
-    @ColumnInfo(name = "created_at") val createdAt: Int,
+    @ColumnInfo(name = "created_at") val createdAt: Long,
 )
